+++ mono/lib/src/commands/task.dart
@@ -1,4 +1,4 @@
-import 'package:mono_cli/mono_cli.dart';
+import 'package:mono_core/mono_core.dart';
 
 class TaskCommand {
   /// Attempts to run an arbitrary task by name as a top-level command.