+++ mono/test/commands/get_command_test.dart
@@ -4,6 +4,7 @@ import 'package:mono/src/commands/get.dart';
 import 'package:mono_cli/mono_cli.dart';
 import 'package:test/test.dart';
 import 'package:path/path.dart' as p;
+import 'package:mono_core/mono_core.dart';
 
 import '../util/fs_fixtures.dart';
 import '../util/fakes.dart';