+++ mono_cli/lib/src/config/yaml_loader.dart
@@ -1,4 +1,5 @@
 import 'package:mono_cli/mono_cli.dart';
+import 'package:mono_core/mono_core.dart';
 
 @immutable
 class YamlConfigLoader implements ConfigLoader {