+++ mono_cli/lib/src/system_io/group_store_fs.dart
@@ -1,4 +1,4 @@
-import 'package:mono_cli/mono_cli.dart';
+import 'package:mono_core/mono_core.dart';
 
 class FileGroupStore implements GroupStore {
   FileGroupStore(this._folder);