+++ mono_cli/lib/src/system_io/path_service.dart
@@ -1,4 +1,4 @@
-import 'package:mono_cli/mono_cli.dart';
+import 'package:mono_core/mono_core.dart';
 import 'package:path/path.dart' as p;
 
 class DefaultPathService implements PathService {