+++ mono_cli/lib/src/system_io/prompt.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:mono_cli/mono_cli.dart';
+import 'package:mono_core/mono_core.dart';
 
 class ConsolePrompter implements Prompter {
   const ConsolePrompter();