+++ mono_cli/lib/src/system_io/workspace_config_fs.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 
 import 'package:mono_cli/mono_cli.dart';
+import 'package:mono_core/mono_core.dart';
 
 class FileWorkspaceConfig implements WorkspaceConfig {
   const FileWorkspaceConfig();